*Se agrega método para calcular tangente inversa en todos los casos

+++ Simulacion/src/plano/Angulo.scala
@@ -1,3 +1,3 @@
 package plano
-case class Angulo (val grados:Int){
+case class Angulo (val grados:Double){
 }
